Changed stats to use a base stat seed
Boss monsters and default weapons are broken.

+++ Globals.cpp
@@ -5,5 +5,5 @@ GameStatus gameStatus;
 int menuCase;
 moveType playerMove;
 moveType enemyMove;
-int turnState;
-int menuNum;
+uint16_t turnState;
+uint16_t menuNum;
